Update main.py
The Translation Z is an essential information for the size of the hand.
The current model does not change the overall size of the hand but assess its "closeness" (the 3d model of the hand will always have the same size, it will simply be further away from the camera point of view in the 3d scene).
For the sake of the example, simply divide greatly tz to see all models in a reasonable 3d space size.